# backend/swagger.yaml
openapi: 3.0.3
info:
  title: BHARATNIVESH API
  description: API documentation for BHARATNIVESH platform (core Phase 0 + Phase 1 stubs)
  version: 0.1.0
servers:
  - url: http://localhost:5000
    description: Local development server
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
  /api/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
      responses:
        '201':
          description: Created
  /api/auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
  /api/auth/me:
    get:
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /api/auth/profile:
    put:
      summary: Update profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                risk_profile:
                  type: string
                mode:
                  type: string
      responses:
        '200':
          description: Updated
  /api/onboarding/kyc:
    post:
      summary: Submit KYC info (mock)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pan:
                  type: string
                aadhar:
                  type: string
                address:
                  type: string
      responses:
        '200':
          description: KYC accepted (mock)
  /api/onboarding/quiz:
    post:
      summary: Submit risk profiling quiz
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                answers:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: Profile saved
  /api/market/indices:
    get:
      summary: Get market indices (NIFTY, SENSEX)
      responses:
        '200':
          description: Indices data
  /api/market/quotes:
    get:
      summary: Get quotes for symbols
      parameters:
        - in: query
          name: symbols
          schema:
            type: string
          description: Comma-separated symbols
      responses:
        '200':
          description: Quotes returned
  /api/ipos:
    get:
      summary: Get upcoming IPOs
      responses:
        '200':
          description: List of IPOs
  /api/ipos/{id}:
    get:
      summary: Get IPO details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: IPO details
  /api/ipos/{id}/verdict:
    get:
      summary: Get verdict and red flags for an IPO
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: mode
          required: false
          schema:
            type: string
            enum: [regular, nerd]
      responses:
        '200':
          description: Verdict and red flags
  /api/stocks/screener:
    get:
      summary: Stock screener
      parameters:
        - in: query
          name: strategy
          schema:
            type: string
            enum: [growth, value, momentum, quality]
        - in: query
          name: sector
          schema:
            type: string
        - in: query
          name: minPe
          schema:
            type: number
        - in: query
          name: maxPe
          schema:
            type: number
        - in: query
          name: minMarketCap
          schema:
            type: number
      responses:
        '200':
          description: Screener results
  /api/stocks/{symbol}:
    get:
      summary: Get stock fundamentals (mock)
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Stock details
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT